requirements.txt
Flask==2.3.2
openai==1.14.3
supabase==1.0.3
postgrest-py==0.10.6
gotrue==1.2.0
requests==2.31.0
Flask-Session>=0.4.0,<0.5
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3